+++ src/main/java/dev/muon/otherworldorigins/entity/summons/SummonedGrizzlyBear.java
@@ -1,0 +1,226 @@
+package dev.muon.otherworldorigins.entity.summons;
+
+
+import com.github.alexthe666.alexsmobs.entity.EntityGrizzlyBear;
+import dev.muon.otherworldorigins.effect.ModEffects;
+import dev.muon.otherworldorigins.entity.ModEntities;
+import dev.muon.otherworldorigins.spells.ModSpells;
+import io.redspace.ironsspellbooks.api.registry.SpellRegistry;
+import io.redspace.ironsspellbooks.api.spells.AbstractSpell;
+import io.redspace.ironsspellbooks.api.util.Utils;
+import io.redspace.ironsspellbooks.capabilities.magic.MagicManager;
+import io.redspace.ironsspellbooks.effect.SummonTimer;
+import io.redspace.ironsspellbooks.entity.mobs.MagicSummon;
+import io.redspace.ironsspellbooks.entity.mobs.goals.*;
+import io.redspace.ironsspellbooks.registries.EntityRegistry;
+import io.redspace.ironsspellbooks.registries.MobEffectRegistry;
+import io.redspace.ironsspellbooks.util.OwnerHelper;
+import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.nbt.CompoundTag;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.ai.attributes.AttributeSupplier;
+import net.minecraft.world.entity.ai.attributes.Attributes;
+import net.minecraft.world.entity.ai.goal.FloatGoal;
+import net.minecraft.world.entity.ai.goal.LookAtPlayerGoal;
+import net.minecraft.world.entity.ai.goal.MeleeAttackGoal;
+import net.minecraft.world.entity.ai.goal.WaterAvoidingRandomStrollGoal;
+import net.minecraft.world.entity.animal.PolarBear;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.Vec3;
+
+import javax.annotation.Nullable;
+import java.util.UUID;
+
+public class SummonedGrizzlyBear extends EntityGrizzlyBear implements MagicSummon {
+    protected LivingEntity cachedSummoner;
+    protected UUID summonerUUID;
+    private final Level level = this.level();
+
+    public SummonedGrizzlyBear(EntityType<? extends EntityGrizzlyBear> pEntityType, Level pLevel) {
+        super(pEntityType, pLevel);
+        this.xpReward = 0;
+    }
+
+    public SummonedGrizzlyBear(Level pLevel, LivingEntity owner) {
+        this(ModEntities.SUMMONED_GRIZZLY_BEAR.get(), pLevel);
+        this.setSummoner(owner);
+    }
+
+    public float getStepHeight() {
+        return 1.0F;
+    }
+
+    public void registerGoals() {
+        this.goalSelector.addGoal(0, new FloatGoal(this));
+        this.goalSelector.addGoal(1, new SummonedGrizzlyBear.GrizzlyBearMeleeAttackGoal());
+        this.goalSelector.addGoal(7, new GenericFollowOwnerGoal(this, this::getSummoner, 0.8999999761581421, 15.0F, 5.0F, false, 25.0F));
+        this.goalSelector.addGoal(8, new WaterAvoidingRandomStrollGoal(this, 0.8));
+        this.goalSelector.addGoal(9, new LookAtPlayerGoal(this, Player.class, 3.0F, 1.0F));
+        this.goalSelector.addGoal(10, new LookAtPlayerGoal(this, Mob.class, 8.0F));
+        this.targetSelector.addGoal(1, new GenericOwnerHurtByTargetGoal(this, this::getSummoner));
+        this.targetSelector.addGoal(2, new GenericOwnerHurtTargetGoal(this, this::getSummoner));
+        this.targetSelector.addGoal(3, new GenericCopyOwnerTargetGoal(this, this::getSummoner));
+        this.targetSelector.addGoal(4, (new GenericHurtByTargetGoal(this, (entity) -> {
+            return entity == this.getSummoner();
+        })).setAlertOthers(new Class[0]));
+    }
+
+    public InteractionResult mobInteract(Player pPlayer, InteractionHand pHand) {
+        if (this.isVehicle()) {
+            return super.mobInteract(pPlayer, pHand);
+        } else {
+            if (pPlayer == this.getSummoner()) {
+                this.doPlayerRide(pPlayer);
+            }
+
+            return InteractionResult.sidedSuccess(this.level.isClientSide);
+        }
+    }
+
+    protected void doPlayerRide(Player pPlayer) {
+        this.setStanding(false);
+        if (!this.level.isClientSide) {
+            pPlayer.setYRot(this.getYRot());
+            pPlayer.setXRot(this.getXRot());
+            pPlayer.startRiding(this);
+        }
+
+    }
+
+    public LivingEntity getSummoner() {
+        return OwnerHelper.getAndCacheOwner(this.level, this.cachedSummoner, this.summonerUUID);
+    }
+
+    public void setSummoner(@Nullable LivingEntity owner) {
+        if (owner != null) {
+            this.summonerUUID = owner.getUUID();
+            this.cachedSummoner = owner;
+        }
+
+    }
+
+    public void die(DamageSource pDamageSource) {
+        this.onDeathHelper();
+        super.die(pDamageSource);
+    }
+
+    public void onRemovedFromWorld() {
+        this.onRemovedHelper(this, ModEffects.BEAST_TIMER.get());
+        super.onRemovedFromWorld();
+    }
+
+    public void readAdditionalSaveData(CompoundTag compoundTag) {
+        super.readAdditionalSaveData(compoundTag);
+        this.summonerUUID = OwnerHelper.deserializeOwner(compoundTag);
+    }
+
+    public void addAdditionalSaveData(CompoundTag compoundTag) {
+        super.addAdditionalSaveData(compoundTag);
+        OwnerHelper.serializeOwner(compoundTag, this.summonerUUID);
+    }
+
+    public boolean doHurtTarget(Entity pEntity) {
+        return Utils.doMeleeAttack(this, pEntity, (ModSpells.SUMMON_GRIZZLY_BEAR.get()).getDamageSource(this, this.getSummoner()));
+    }
+
+    public boolean isAlliedTo(Entity pEntity) {
+        return super.isAlliedTo(pEntity) || this.isAlliedHelper(pEntity);
+    }
+
+    public void onUnSummon() {
+        if (!this.level.isClientSide) {
+            MagicManager.spawnParticles(this.level, ParticleTypes.POOF, this.getX(), this.getY(), this.getZ(), 25, 0.4, 0.8, 0.4, 0.03, false);
+            this.discard();
+        }
+
+    }
+
+    public boolean hurt(DamageSource pSource, float pAmount) {
+        return this.shouldIgnoreDamage(pSource) ? false : super.hurt(pSource, pAmount);
+    }
+
+    public static AttributeSupplier.Builder createAttributes() {
+        return Mob.createMobAttributes().add(Attributes.MAX_HEALTH, 30.0).add(Attributes.FOLLOW_RANGE, 20.0).add(Attributes.MOVEMENT_SPEED, 0.3).add(Attributes.ATTACK_DAMAGE, 6.0);
+    }
+
+    @Nullable
+    public LivingEntity getControllingPassenger() {
+        Entity entity = this.getFirstPassenger();
+        if (entity instanceof Mob) {
+            return (Mob)entity;
+        } else {
+            entity = this.getFirstPassenger();
+            return entity instanceof Player ? (Player)entity : null;
+        }
+    }
+
+    protected void tickRidden(Player player, Vec3 p_275242_) {
+        super.tickRidden(player, p_275242_);
+        this.yRotO = this.getYRot();
+        this.setYRot(player.getYRot());
+        this.setXRot(player.getXRot());
+        this.setRot(this.getYRot(), this.getXRot());
+        this.yBodyRot = this.yRotO;
+        this.yHeadRot = this.getYRot();
+    }
+
+    protected Vec3 getRiddenInput(Player player, Vec3 p_275300_) {
+        float f = player.xxa * 0.5F;
+        float f1 = player.zza;
+        if (f1 <= 0.0F) {
+            f1 *= 0.25F;
+        }
+
+        if (this.isInWater()) {
+            f *= 0.3F;
+            f1 *= 0.3F;
+        }
+
+        return new Vec3((double)f, 0.0, (double)f1);
+    }
+
+    protected float getRiddenSpeed(Player p_278336_) {
+        return (float)this.getAttributeValue(Attributes.MOVEMENT_SPEED) * 0.8F;
+    }
+
+    class GrizzlyBearMeleeAttackGoal extends MeleeAttackGoal {
+        public GrizzlyBearMeleeAttackGoal() {
+            super(SummonedGrizzlyBear.this, 1.25, true);
+        }
+
+        protected void checkAndPerformAttack(LivingEntity pEnemy, double pDistToEnemySqr) {
+            double d0 = this.getAttackReachSqr(pEnemy);
+            if (pDistToEnemySqr <= d0 && this.isTimeToAttack()) {
+                this.resetAttackCooldown();
+                this.mob.doHurtTarget(pEnemy);
+                SummonedGrizzlyBear.this.setStanding(false);
+            } else if (pDistToEnemySqr <= d0 * 2.0) {
+                if (this.isTimeToAttack()) {
+                    SummonedGrizzlyBear.this.setStanding(false);
+                    this.resetAttackCooldown();
+                }
+
+                if (this.getTicksUntilNextAttack() <= 10) {
+                    SummonedGrizzlyBear.this.setStanding(true);
+                }
+            } else {
+                this.resetAttackCooldown();
+                SummonedGrizzlyBear.this.setStanding(false);
+            }
+
+        }
+
+        public void stop() {
+            SummonedGrizzlyBear.this.setStanding(false);
+            super.stop();
+        }
+    }
+}
+