+++ src/main/java/dev/muon/otherworldorigins/spells/SummonGrizzlyBearSpell.java
@@ -39,6 +39,21 @@ public class SummonGrizzlyBearSpell extends AbstractSpell {
         this.baseManaCost = 50;
     }
 
+    @Override
+    public boolean isEnabled() {
+        return false;
+    }
+
+    @Override
+    public boolean allowCrafting() {
+        return false;
+    }
+
+    @Override
+    public boolean allowLooting() {
+        return false;
+    }
+
     public CastType getCastType() {
         return CastType.LONG;
     }